[modified]inner camera with barcode
